requirements.txt
absl-py==2.3.1
agent-starter-pack==0.26.2
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.17.2
annotated-types==0.7.0
anyio==4.12.0
arrow==1.4.0
asn1crypto==1.5.1
attrs==25.4.0
authlib==1.6.5
backoff==2.2.1
bcrypt==4.0.1
binaryornot==0.4.4
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
cloud-sql-python-connector==1.19.0
cloudpickle==3.1.2
codespell==2.4.1
cookiecutter==2.6.0
cryptography==46.0.3
dnspython==2.8.0
docstring-parser==0.17.0
ecdsa==0.19.1
email-validator==2.3.0
et-xmlfile==2.0.0
fastapi==0.118.3
-e file:///Users/sreenikethaathreya/adksolutionsaccelerator
frozenlist==1.8.0
google-adk==1.20.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.129.0
google-cloud-appengine-logging==1.7.0
google-cloud-audit-log==0.4.0
google-cloud-bigquery==3.38.0
google-cloud-bigquery-storage==2.35.0
google-cloud-bigtable==2.34.0
google-cloud-core==2.5.0
google-cloud-discoveryengine==0.13.12
google-cloud-firestore==2.21.0
google-cloud-logging==3.12.1
google-cloud-monitoring==2.28.0
google-cloud-resource-manager==1.15.0
google-cloud-secret-manager==2.25.0
google-cloud-spanner==3.59.0
google-cloud-speech==2.34.0
google-cloud-storage==3.7.0
google-cloud-trace==1.17.0
google-crc32c==1.7.1
google-genai==1.54.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
graphviz==0.21
grpc-google-iam-v1==0.14.3
grpc-interceptor==0.15.4
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib-metadata==8.7.0
iniconfig==2.3.0
jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
librt==0.7.3
mako==1.3.10
markdown-it-py==4.0.0
markupsafe==3.0.3
mcp==1.23.3
mdurl==0.1.2
multidict==6.7.0
mypy==1.19.0
mypy-extensions==1.1.0
numpy==2.3.5
oauthlib==3.3.1
openpyxl==3.1.5
opentelemetry-api==1.37.0
opentelemetry-exporter-gcp-logging==1.11.0a0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-google-genai==0.4b0
opentelemetry-proto==1.37.0
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-genai==0.2b0
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pathspec==0.12.1
pg8000==1.31.5
pillow==12.0.0
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.2
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pygments==2.19.2
pyjwt==2.10.1
pyparsing==3.2.5
pypdf2==3.0.1
pytesseract==0.3.13
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
pyyaml==6.0.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
ruff==0.14.8
scramp==1.4.6
shapely==2.1.2
six==1.17.0
sqlalchemy==2.0.45
sqlalchemy-spanner==1.17.1
sqlparse==0.5.4
sse-starlette==3.0.3
starlette==0.48.0
tabulate==0.9.0
tenacity==9.1.2
text-unidecode==1.3
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.1
uvicorn==0.38.0
uvloop==0.22.1
vertexai==1.71.1
watchdog==6.0.0
watchfiles==1.1.1
websockets==15.0.1
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
